requirements.txt
streamlit==1.35.0
pandas==2.2.2
tiktoken==0.7.0
langchain==0.2.3
python-docx==1.1.2
